renamed widget, modified README wording
