+++ src/common/example/example.css
@@ -16,7 +16,7 @@
 @import '../../tabcontroller/styles/tabController.m.css';
 @import '../../textarea/styles/textarea.m.css';
 @import '../../textinput/styles/textinput.m.css';
-@import '../../decoratedtextinput/styles/decoratedtextinput.m.css';
+@import '../../enhancedtextinput/styles/enhancedtextinput.m.css';
 @import '../../timepicker/styles/timePicker.m.css';
 @import '../../titlepane/styles/titlePane.m.css';
 @import '../../accordionpane/styles/accordionPane.m.css';
