+++ src/decoratedtextinput/EnhancedTextInput.ts
@@ -3,15 +3,15 @@ import { theme } from '@dojo/widget-core/mixins/Themed';
 import { v } from '@dojo/widget-core/d';
 
 import TextInput, { TextInputProperties } from '../textinput/TextInput';
-import * as css from './styles/decoratedtextinput.m.css';
+import * as css from './styles/enhancedtextinput.m.css';
 
-export interface DecoratedTextInputProperties extends TextInputProperties {
-	addonAfter: DNode[];
-	addonBefore: DNode[];
+export interface EnhancedTextInputProperties extends TextInputProperties {
+	addonAfter?: DNode[];
+	addonBefore?: DNode[];
 }
 
 @theme(css)
-export default class DecoratedTextInput extends TextInput<DecoratedTextInputProperties> {
+export default class EnhancedTextInput extends TextInput<EnhancedTextInputProperties> {
 	protected renderAddon(addon: DNode, before = false) {
 		return v('span', {
 			classes: this.theme([css.addon, before ? css.addonBefore : css.addonAfter])
