+++ src/themes/dojo/widgets.css
@@ -3,7 +3,7 @@
 @import './calendar.m.css';
 @import './checkbox.m.css';
 @import './comboBox.m.css';
-@import './decoratedtextinput.m.css';
+@import './enhancedtextinput.m.css';
 @import './dialog.m.css';
 @import './icons.m.css';
 @import './label.m.css';
